Use typer to reformat the CLI functionality of the app
